Move SearchReferences base controller under Synonyms namespace
Because it is not being used outside it, so makes sense for it to live
there
